split interval mergers into different file
switch on memory markup with special service debug mode
