+++ ydb/core/tx/columnshard/engines/reader/plain_reader/iterator/ya.make
@@ -7,6 +7,7 @@ SRCS(
     interval.cpp
     fetched_data.cpp
     plain_read_data.cpp
+    merge.cpp
     columns_set.cpp
     context.cpp
     fetching.cpp
