+++ ydb/library/services/services.proto
@@ -73,6 +73,7 @@ enum EServiceKikimr {
     TX_COLUMNSHARD = 332;
     BLOB_CACHE = 333;
     TX_COLUMNSHARD_SCAN = 334;
+    TX_COLUMNSHARD_SCAN_MEMORY = 335;
 
     // BLOBSTORAGE again
     BS_HANDOFF = 298;
